python-ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
SpeechRecognition==3.14.3
pydub==0.25.1
audioop-lts==0.2.2
werkzeug==3.0.1
sentence-transformers
faiss-cpu
numpy
scikit-learn
Pillow
pytesseract
transformers
torch
requests
beautifulsoup4
lxml
rapidfuzz
langdetect
